requirements.txt
alembic==1.9.2
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
crypto==1.4.1
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.2
Naked==0.1.32
oauth2==1.9.0.post1
orjson==3.8.5
passlib==1.7.4
psycopg2==2.9.5
pycryptodome==3.17
pydantic==1.10.4
pyparsing==3.0.9
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.1
starlette==0.22.0
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
